Experiment to use "has_thing" within a macro
Defining the "has_thing" operation within a macro definition results in
a compiler error when used with classes without the specified member.
Really not sure why, may have something to do with the containing
function not being a template function itself, so the compiler is unable
to properly resolve the constexpr.