Use vertex shader to draw background
Use a vertex shader that draws a single triangle covering the full
window to draw the background.

By eliminating the need to use a quad to draw the background texture,
this shader also eliminates the need to allocate and bind the vertex
buffer for said quad.

Some background (and possibly the OG implementation) here:
https://web.archive.org/web/20140719063725/http://www.altdev.co/2011/08/08/interesting-vertex-shader-trick/